refactor: Rename package, RandomMovableStrategy
- 패키지명 수정
- bound값 RandomMovableStrategy 내부 이동

(cherry picked from commit fbd9477014234d7bf02750f7cab5d50346385aa8)
