+++ src/main/java/racingcar/domain/movable/RandomMovableStrategy.java
@@ -6,15 +6,14 @@ public class RandomMovableStrategy implements MovableStrategy {
 
     private final Random RANDOM = new Random();
     private final int CONDITION_MOVE_AVAILABLE = 4;
-    private final int bound;
+    private final int DEFAULT_BOUND = 10;
 
-    public RandomMovableStrategy(final int bound) {
-        this.bound = bound;
+    public RandomMovableStrategy() {
     }
 
     @Override
     public boolean isMovable() {
-        return getRandom(bound) >= CONDITION_MOVE_AVAILABLE;
+        return getRandom(DEFAULT_BOUND) >= CONDITION_MOVE_AVAILABLE;
     }
 
     private int getRandom(final int bound) {
