+++ src/test/java/racingcar/TestUtils.java
@@ -6,13 +6,11 @@ import racingcar.domain.movable.RandomMovableStrategy;
 
 public class TestUtils {
 
-    private static final int RANDOM_BOUND = 10;
-
     private TestUtils() {}
 
     static void carMove(Car car, List<Boolean> moves) {
         for (boolean carMoved: moves) {
-            car.run(new RandomMovableStrategy(RANDOM_BOUND) {
+            car.run(new RandomMovableStrategy() {
                 @Override
                 public boolean isMovable() {
                     return carMoved;
